+++ include/allocators/FixedBlockAllocator.h
@@ -37,7 +37,7 @@ namespace BlackMagic
 				storage = (FixedBlockTracking*)buffer;
 			}
 
-			byte alignmentIssues = (reinterpret_cast<size_t>(storage) & (alignment - 1));
+			byte alignmentIssues = (byte)(reinterpret_cast<size_t>(storage) & (alignment - 1));
 			size -= alignmentIssues;
 			storage = reinterpret_cast<FixedBlockTracking*>(reinterpret_cast<byte*>(storage) + alignmentIssues);
